functions-python/requirements.txt
# Python Cloud Function requirements for image generation
# Using Hugging Face Inference API (no model download needed!)

# Firebase Functions SDK (required for Python Cloud Functions)
firebase-functions>=0.1.0
firebase-admin>=6.0.0

# HTTP requests for Hugging Face API
requests>=2.31.0

# Image processing
pillow>=9.0.0

# Google Cloud
google-cloud-storage>=2.10.0
google-cloud-firestore>=2.11.0

